refactor(client): simplify PNG export by removing ExportProvider and ExportRegistrar
