--- client/src/lib/webshot.ts
@@ -1,78 +1,0 @@
-import { MutationFunction, useMutation, UseMutationOptions } from "@tanstack/react-query";
-import axios from "axios";
-
-/**
- ************************************************************
- ************************************************************
- * WebshotReport
- * usePostWebshotReportMutation
- ************************************************************
- ************************************************************
- */
-
-export type PostWebshotReportParams = {
-  pagePath: string;
-  generatedTextContent?: Record<string, unknown>;
-};
-
-export const postWebshotReport = async (payload: PostWebshotReportParams) => {
-  return axios.request<Blob>({
-    url: "/local-api/webshot/report",
-    method: "POST",
-    responseType: "blob",
-    headers: { "Content-Type": "application/json" },
-    data: payload,
-  });
-};
-
-export const postWebshotReportMutationOptions = <TError = unknown, TContext = unknown>(
-  options?: UseMutationOptions<
-    Awaited<ReturnType<typeof postWebshotReport>>,
-    TError,
-    PostWebshotReportParams,
-    TContext
-  >,
-): UseMutationOptions<
-  Awaited<ReturnType<typeof postWebshotReport>>,
-  TError,
-  PostWebshotReportParams,
-  TContext
-> => {
-  const mutationFn: MutationFunction<
-    Awaited<ReturnType<typeof postWebshotReport>>,
-    PostWebshotReportParams
-  > = (vars) => postWebshotReport(vars);
-  return { mutationFn, ...options };
-};
-
-export const usePostWebshotReportMutation = <TError = unknown, TContext = unknown>(
-  options?: UseMutationOptions<
-    Awaited<ReturnType<typeof postWebshotReport>>,
-    TError,
-    PostWebshotReportParams,
-    TContext
-  >,
-) => {
-  return useMutation(postWebshotReportMutationOptions(options));
-};
-
-export async function downloadBlobResponse(blob: Blob, filename: string) {
-  if (!blob || !filename) return;
-
-  const url = URL.createObjectURL(blob);
-  const a = document.createElement("a");
-  a.href = url;
-  a.download = decodeURIComponent(filename.trim());
-  document.body.appendChild(a);
-  a.click();
-  a.remove();
-  URL.revokeObjectURL(url);
-}
-
-const date = new Date().toISOString().split("T")[0];
-
-export function filenameFromCD(cd?: string | null, fallback = `report_${date}.pdf`) {
-  if (!cd) return fallback;
-  const m = /filename\*?=.*?''?([^;]+)|filename="?([^"]+)"?/i.exec(cd);
-  return decodeURIComponent((m?.[1] || m?.[2] || fallback).trim());
-}
