+++ client/src/lib/webshot/index.ts
@@ -1,0 +1,3 @@
+export { downloadBlobResponse } from "./download";
+export { exportToPng } from "./export-to-png";
+export { registerMapForExport, unregisterMapForExport } from "./map-export-registry";
